5_RAG_database/requirements.txt
torch
transformers
pillow
faiss-cpu
openai
numpy
